Don't wipe someones odasrvmgr.toml if setup.sh is rerun

+++ setup.sh
@@ -54,7 +54,9 @@ install -T -m 644  -o root -g root "$bash_completion_src" "/usr/share/bash-compl
 install -T -D -m 644 -o root -g root "$script_dir/odasrvargs.sh" "$install_dir/bin/odasrvargs.sh"
 install -T -D -m 644 -o root -g root "$script_dir/tomlconfig.py" "$install_dir/bin/tomlconfig.py"
 install -T -D -m 644 -o root -g root "$script_dir/wadfetch.py" "$install_dir/bin/wadfetch.py"
-install -T -D -m 664 -o root -g odasrvmgr "$script_dir/odasrvmgr.toml" "/etc/odasrvmgr/odasrvmgr.toml"
+if [[ ! -f "/etc/odasrvmgr/odasrvmgr.toml" ]]; then
+  install -T -D -m 664 -o root -g odasrvmgr "$script_dir/odasrvmgr.toml" "/etc/odasrvmgr/odasrvmgr.toml"
+fi
 
 # Enable systemd services
 echo "Enabling servers..."
